Refactor for API and events
