+++ src/API/Common/CarrySlot.cs
@@ -1,4 +1,4 @@
-namespace CarryOn
+namespace CarryOn.API.Common
 {
     /// <summary> Represents a location where something is or may be carried. </summary>
     public enum CarrySlot
