+++ src/API/Common/ICarryEvent.cs
@@ -1,0 +1,7 @@
+namespace CarryOn.API.Common
+{
+    public interface ICarryEvent
+    {
+      void Init(CarrySystem carrySystem);
+    }
+}
