+++ src/API/Event/BlockEntityDataDelegate.cs
@@ -1,0 +1,7 @@
+using Vintagestory.API.Common;
+using Vintagestory.API.Datastructures;
+
+namespace CarryOn.API.Event
+{
+    public delegate void BlockEntityDataDelegate(BlockEntity blockEntity, ITreeAttribute blockEntityData);
+}
