+++ src/API/Event/BlockRemovedEventArgs.cs
@@ -1,0 +1,15 @@
+using System;
+using CarryOn.API.Common;
+using Vintagestory.API.Common;
+using Vintagestory.API.Common.Entities;
+using Vintagestory.API.MathTools;
+
+namespace CarryOn.API.Event
+{
+    public class BlockRemovedEventArgs : EventArgs
+    {
+        public IWorldAccessor World { get; set; }
+
+        public BlockPos Position  {get;set;}
+    }
+}
