+++ src/Client/AnimationFixer.cs
@@ -1,7 +1,7 @@
 using System.Collections.Generic;
 using System.Linq;
+using CarryOn.API.Common;
 using Vintagestory.API.Common;
-
 namespace CarryOn.Client
 {
     /// <summary> Utility class which is meant to syncronize animations for
