+++ src/Client/EntityCarryRenderer.cs
@@ -1,6 +1,7 @@
 using System;
 using System.Collections.Generic;
 using System.Linq;
+using CarryOn.API.Common;
 using Vintagestory.API.Client;
 using Vintagestory.API.Common;
 using Vintagestory.API.MathTools;
@@ -47,9 +48,7 @@ namespace CarryOn.Client
 
             var renderInfo = Api.Render.GetItemStackRenderInfo(new DummySlot(carried.ItemStack), EnumItemRenderTarget.Ground);
 
-            var behavior = carried.Behavior;
-
-            renderInfo.Transform = behavior.Slots[carried.Slot]?.Transform ?? behavior.DefaultTransform;
+            renderInfo.Transform = carried.Behavior.Slots[carried.Slot]?.Transform ?? carried.Behavior.DefaultTransform;
             return renderInfo;
         }
