+++ src/Server/EntityBehaviorDropCarriedOnDamage.cs
@@ -1,3 +1,4 @@
+using CarryOn.API.Common;
 using Vintagestory.API.Common;
 using Vintagestory.API.Common.Entities;
